acme.sh自动下载新版，zerossl支持，renew支持，cloudflare支持，代理支持，证书切换到ECC

+++ cert-up.sh
@@ -15,6 +15,8 @@ CRT_PATH=${CRT_BASE_PATH}/_archive/${CRT_PATH_NAME}
 FIND_MAJORVERSION_FILE="/etc/VERSION"
 FIND_MAJORVERSION_STR="majorversion=\"7\""
 
+versionLt () { test "$(echo "$@" | tr " " "\n" | sort -rV | head -n 1)" != "$1"; }
+
 backupCrt () {
   echo 'begin backupCrt'
   BACKUP_PATH=${BASE_ROOT}/backup/${DATE_TIME}
@@ -27,31 +29,57 @@ backupCrt () {
 }
 
 installAcme () {
-  echo 'begin installAcme'
-  mkdir -p ${TEMP_PATH}
-  cd ${TEMP_PATH}
-  echo 'begin downloading acme.sh tool...'
-  ACME_SH_ADDRESS=`curl -L https://cdn.jsdelivr.net/gh/andyzhshg/syno-acme@master/acme.sh.address`
-  SRC_TAR_NAME=acme.sh.tar.gz
-  curl -L -o ${SRC_TAR_NAME} ${ACME_SH_ADDRESS}
-  SRC_NAME=`tar -tzf ${SRC_TAR_NAME} | head -1 | cut -f1 -d"/"`
-  tar zxvf ${SRC_TAR_NAME}
-  echo 'begin installing acme.sh tool...'
-  cd ${SRC_NAME}
-  ./acme.sh --install --nocron --home ${ACME_BIN_PATH}
-  echo 'done installAcme'
-  rm -rf ${TEMP_PATH}
+  ALLOW_INSTALL=false
+  ACME_SH_FILE=${ACME_BIN_PATH}/acme.sh
+  ACME_SH_NEW_VERSION=$(wget -qO- -t1 -T2 "https://api.github.com/repos/acmesh-official/acme.sh/releases/latest" | grep "tag_name" | head -n 1 | awk -F ":" '{print $2}' | sed 's/\"//g;s/,//g;s/ //g')
+  ACME_SH_ADDRESS=https://ghproxy.com/https://github.com/acmesh-official/acme.sh/archive/${ACME_SH_NEW_VERSION}.tar.gz
+  if [ -z "${ACME_SH_NEW_VERSION}" ]; then
+    echo 'unable to get new version number'
+    return 0
+  fi
+  if [ ! -f "${ACME_SH_FILE}" ]; then
+    ALLOW_INSTALL=true
+    echo 'acme not installed, start install'
+  else
+    ACME_SH_VERSION=$(cat ${ACME_SH_FILE} | grep "VER=*" | head -n 1 | awk -F "=" '{print $2}' | sed 's/\"//g;s/,//g;s/ //g')
+    if versionLt ${ACME_SH_VERSION} ${ACME_SH_NEW_VERSION}; then
+      ALLOW_INSTALL=true
+      echo 'acme has a new version, start updating'
+    else
+      echo 'skip acme installation'
+    fi
+  fi
+  if [ ${ALLOW_INSTALL} == true ]; then
+    echo 'in progress...'
+    mkdir -p ${TEMP_PATH}
+    cd ${TEMP_PATH}
+    echo 'begin downloading acme.sh tool...'
+    SRC_TAR_NAME=acme.sh.tar.gz
+    curl -L -o ${SRC_TAR_NAME} ${ACME_SH_ADDRESS}
+    SRC_NAME=`tar -tzf ${SRC_TAR_NAME} | head -1 | cut -f1 -d"/"`
+    tar zxvf ${SRC_TAR_NAME}
+    echo 'begin installing acme.sh tool...'
+    cd ${SRC_NAME}
+    ./acme.sh --install --nocron --home ${ACME_BIN_PATH}
+    echo 'done installAcme'
+    rm -rf ${TEMP_PATH}
+  fi
   return 0
 }
 
 generateCrt () {
   echo 'begin generateCrt'
   cd ${BASE_ROOT}
-  source config
+  source ./config
+  # add register zerossl account
+  if [ ${CERT_SERVER} == 'zerossl' ]; then
+    echo 'register zerossl account'
+    ${ACME_BIN_PATH}/acme.sh  --register-account  -m ${ACCOUNT_EMAIL} --server zerossl
+  fi
   echo 'begin updating default cert by acme.sh tool'
   source ${ACME_BIN_PATH}/acme.sh.env
-  ${ACME_BIN_PATH}/acme.sh --force --log --issue --dns ${DNS} --dnssleep ${DNS_SLEEP} -d "${DOMAIN}" -d "*.${DOMAIN}"
-  ${ACME_BIN_PATH}/acme.sh --force --installcert -d ${DOMAIN} -d *.${DOMAIN} \
+  ${ACME_BIN_PATH}/acme.sh --force --log --issue --server ${CERT_SERVER} --dns ${DNS} --dnssleep ${DNS_SLEEP} -d "${DOMAIN}" -d "*.${DOMAIN}" --keylength ec-256
+  ${ACME_BIN_PATH}/acme.sh --force --installcert -d ${DOMAIN} -d *.${DOMAIN} --ecc \
     --certpath ${CRT_PATH}/cert.pem \
     --key-file ${CRT_PATH}/privkey.pem \
     --fullchain-file ${CRT_PATH}/fullchain.pem
@@ -67,12 +95,34 @@ generateCrt () {
   fi
 }
 
+renewCrt () {
+  echo 'begin renewCrt'
+  cd ${BASE_ROOT}
+  source ./config
+  echo 'begin renew default cert by acme.sh tool'
+  source ${ACME_BIN_PATH}/acme.sh.env
+  ${ACME_BIN_PATH}/acme.sh --cron --home ${ACME_BIN_PATH}
+  ${ACME_BIN_PATH}/acme.sh --force --installcert -d ${DOMAIN} -d *.${DOMAIN} --ecc \
+    --certpath ${CRT_PATH}/cert.pem \
+    --key-file ${CRT_PATH}/privkey.pem \
+    --fullchain-file ${CRT_PATH}/fullchain.pem
+  if [ -s "${CRT_PATH}/cert.pem" ]; then
+    echo 'done generateCrt'
+    return 0
+  else
+    echo '[ERR] fail to generateCrt'
+    echo "begin revert"
+    revertCrt
+    exit 1;
+  fi
+}
+
 updateService () {
   echo 'begin updateService'
   echo 'cp cert path to des'
   if [ `grep -c "$FIND_MAJORVERSION_STR" $FIND_MAJORVERSION_FILE` -ne '0' ];then
-    echo "MajorVersion = 7, use system default python2"
-    python2 ${BASE_ROOT}/crt_cp.py ${CRT_PATH_NAME}
+    echo "MajorVersion = 7, use system default python"
+    python ${BASE_ROOT}/crt_cp.py ${CRT_PATH_NAME}
   else
     echo "MajorVersion < 7"
     /bin/python2 ${BASE_ROOT}/crt_cp.py ${CRT_PATH_NAME}
@@ -119,20 +169,34 @@ revertCrt () {
   echo 'done revertCrt'
 }
 
-updateCrt () {
-  echo '------ begin updateCrt ------'
+initGenerateCrt () {
+  echo '------ begin init and generateCrt ------'
   backupCrt
   installAcme
   generateCrt
   updateService
   reloadWebService
-  echo '------ end updateCrt ------'
+  echo '------ end init and generateCrt ------'
+}
+
+cronRenewCrt () {
+  echo '------ begin cronRenewCrt ------'
+  backupCrt
+  renewCrt
+  updateService
+  reloadWebService
+  echo '------ end cronRenewCrt ------'
 }
 
 case "$1" in
-  update)
-    echo "begin update cert"
-    updateCrt
+  generate)
+    echo "begin generate cert"
+    initGenerateCrt
+    ;;
+
+  cron)
+    echo "begin cron renew cert"
+    cronRenewCrt
     ;;
 
   revert)
